Refactor engine includes: break cycles, add core.hpp/all.hpp split
Break circular includes and make the include graph modular:

- Remove circular `#include "timeless/timeless.hpp"` from rendering_system.hpp
  and graph.hpp (neither used any TE:: symbol); no engine header includes the
  umbrella anymore.
- Make the TE:: facade self-sufficient: include entity/settings/animation and
  the glm headers it actually references instead of relying on transitive
  includes. Make the umbrella's re-exports explicit (Text/Model/Sprite/etc.
  were only leaking in via animation.hpp).
- Split into core.hpp (facade + direct deps; avoids the 9 heavy systems incl.
  rendering/sound) and all.hpp (full umbrella). timeless.hpp now forwards to
  all.hpp for backward compatibility.
- Trim transform_system.hpp (empty class) down to system.hpp.
- Fix settings.hpp: add missing #pragma once (caused redefinition errors once
  reached via two include paths) and missing <glm/glm.hpp> for glm::vec4.

Verified with a full WASM build of BiasBusters.
