+++ include/timeless/systems/transform_system.hpp
@@ -1,13 +1,5 @@
 #pragma once
-#include "timeless/components/animation.hpp"
-#include "timeless/components/quad.hpp"
-#include "timeless/components/shader.hpp"
-#include "timeless/components/transform.hpp"
-#include "timeless/managers/component_manager.hpp"
-#include "timeless/systems/system.hpp"
-#include "timeless/timeless.hpp"
-#include <memory>
-#include <string.h>
+#include "timeless/systems/system.hpp" // System base class
 
 class TransformSystem : public System {
 };
